+++ src-ui/app/services/discord.service.ts
@@ -1,23 +1,59 @@
 import { Injectable } from '@angular/core';
 import { SleepService } from './sleep.service';
-import { asyncScheduler, distinctUntilChanged, throttleTime } from 'rxjs';
+import {
+  asyncScheduler,
+  combineLatest,
+  debounceTime,
+  distinctUntilChanged,
+  map,
+  throttleTime,
+} from 'rxjs';
 import { invoke } from '@tauri-apps/api';
 import { TranslateService } from '@ngx-translate/core';
+import { VRChatService } from './vrchat.service';
+import { AppSettingsService } from './app-settings.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class DiscordService {
-  constructor(private sleepService: SleepService, private translate: TranslateService) {}
+  private activityCleared = false;
+
+  constructor(
+    private sleepService: SleepService,
+    private translate: TranslateService,
+    private vrchat: VRChatService,
+    private appSettingsService: AppSettingsService
+  ) {}
 
   async init() {
-    this.sleepService.mode
+    combineLatest([
+      this.appSettingsService.settings.pipe(
+        map((settings) => settings.discordActivityMode),
+        distinctUntilChanged()
+      ),
+      this.appSettingsService.settings.pipe(
+        map((settings) => settings.discordActivityOnlyWhileVRChatIsRunning),
+        distinctUntilChanged()
+      ),
+      this.vrchat.vrchatProcessActive.pipe(distinctUntilChanged()),
+      this.sleepService.mode.pipe(distinctUntilChanged()),
+    ])
       .pipe(
+        debounceTime(100),
+        map(([activityMode, onlyWhileVRChatIsRunning, vrchatActive, sleepMode]) => {
+          if (activityMode === 'DISABLED') return null;
+          if (activityMode === 'ONLY_ASLEEP' && !sleepMode) return null;
+          if (onlyWhileVRChatIsRunning && !vrchatActive) return null;
+          return sleepMode;
+        }),
         distinctUntilChanged(),
         throttleTime(4000, asyncScheduler, { leading: true, trailing: true })
       )
-      .subscribe((sleepMode) => {
-        this.updateActivity(sleepMode);
+      .subscribe(async (sleepMode: boolean | null) => {
+        if (sleepMode === null) {
+          if (!this.activityCleared) await this.clearActivity();
+        } else await this.updateActivity(sleepMode);
       });
   }
 
@@ -30,5 +66,11 @@ export class DiscordService {
       ),
       asset: sleepMode ? 'asleep' : 'awake',
     });
+    this.activityCleared = false;
+  }
+
+  private async clearActivity() {
+    await invoke('discord_clear_activity');
+    this.activityCleared = true;
   }
 }